create class for convert_energy_consumers_cim16
